Restructure css to module.css for Add Data page and components
