--- src/components/AddData/ChooseSurvey.jsx
@@ -1,17 +1,0 @@
-// import { React } from 'react';
-// import { Button } from 'antd';
-// import { FileAddOutlined } from '@ant-design/icons';
-
-// function ChooseSurvey() {
-//   return (
-//     <div className="choose-survey-page">
-//       Choose Survey
-//       <Button type="primary" icon={<FileAddOutlined />}>
-//         Add to existing survey
-//       </Button>
-//       <Button type="primary">Create new survey</Button>
-//     </div>
-//   );
-// }
-
-// export default ChooseSurvey;
