+++ src/components/AddData/ImportCSV/ImportCSV.module.css
@@ -1,0 +1,7 @@
+.addDataDiv {
+  text-align: center;
+  padding: 0.5% 3%;
+  align-items: center;
+  justify-content: center;
+  margin-top: -1%;
+}
