+++ src/components/AddData/ReviewForm/ReviewForm.module.css
@@ -1,0 +1,7 @@
+.reviewForm {
+  padding: 0 3% 3%;
+}
+
+#backNextButtons {
+  text-align: center;
+}
