+++ src/components/AddData/StepsBar/StepsBar.jsx
