+++ src/components/AddData/UploadComplete/UploadComplete.module.css
@@ -1,0 +1,7 @@
+.uploadComplete {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  text-align: center;
+  min-height: 60vh;
+}
